+++ relay-server/src/services/upload.rs
@@ -1,18 +1,17 @@
 //! Service that uploads attachments.
-use std::sync::Arc;
 use std::time::Duration;
 
 use bytes::Bytes;
 use futures::future::BoxFuture;
 use futures::stream::FuturesUnordered;
 use futures::{FutureExt, StreamExt};
+use objectstore_client::{Client, ExpirationPolicy, Session, Usecase};
 use relay_config::UploadServiceConfig;
 use relay_quotas::DataCategory;
 use relay_system::{FromMessage, Interface, NoResponse, Receiver, Recipient, Service};
 use smallvec::smallvec;
-use tokio::sync::Notify;
-use uuid::Uuid;
 
+use crate::constants::DEFAULT_ATTACHMENT_RETENTION;
 use crate::managed::{Counted, Managed, OutcomeError, Quantities, Rejected};
 use crate::services::outcome::DiscardReason;
 use crate::statsd::{RelayCounters, RelayGauges};
@@ -37,7 +36,7 @@ impl Interface for Upload {}
 impl FromMessage<UploadAttachment> for Upload {
     type Response = NoResponse;
 
-    fn from_message(message: UploadAttachment, sender: ()) -> Self {
+    fn from_message(message: UploadAttachment, _sender: ()) -> Self {
         Self::Attachment(message)
     }
 }
@@ -48,7 +47,7 @@ pub struct Attachment {
     /// Attachment metadata.
     pub meta: AttachmentMeta,
     /// The attachment body.
-    payload: Bytes,
+    pub payload: Bytes,
 }
 
 impl Counted for Attachment {
@@ -66,7 +65,7 @@ impl Counted for Attachment {
 /// attachment.
 #[derive(Clone, Debug)]
 pub struct UploadedAttachment {
-    meta: AttachmentMeta,
+    pub stored_id: String,
     uploaded_bytes: usize,
 }
 
@@ -82,8 +81,16 @@ impl Counted for UploadedAttachment {
 /// Metadata of the attachment (stub).
 #[derive(Clone, Debug)]
 pub struct AttachmentMeta {
-    attachment_id: Option<String>,
+    pub attachment_id: Option<String>,
     // TODO: more fields
+    pub scope: AttachmentScope,
+}
+
+/// The attachment scope.
+#[derive(Clone, Debug)]
+pub struct AttachmentScope {
+    pub organization_id: u64,
+    pub project_id: u64,
 }
 
 /// Errors that can occur when trying to upload an attachment.
@@ -121,19 +128,34 @@ pub struct UploadService {
     pending_requests: FuturesUnordered<BoxFuture<'static, Result<(), Rejected<Error>>>>,
     max_concurrent_requests: usize,
     timeout: Duration,
+
+    objectstore_client: Client,
+    attachments_usecase: Usecase,
 }
 
 impl UploadService {
-    pub fn new(config: &UploadServiceConfig) -> Self {
+    pub fn new(config: &UploadServiceConfig) -> anyhow::Result<Option<Self>> {
         let UploadServiceConfig {
+            objectstore_url,
             max_concurrent_requests,
             timeout,
         } = config;
-        Self {
+        let Some(objectstore_url) = objectstore_url else {
+            return Ok(None);
+        };
+
+        let objectstore_client = Client::builder(objectstore_url).build()?;
+        let attachments_usecase = Usecase::new("attachments")
+            .with_expiration(ExpirationPolicy::TimeToLive(DEFAULT_ATTACHMENT_RETENTION));
+
+        Ok(Some(Self {
             pending_requests: FuturesUnordered::new(),
             max_concurrent_requests: *max_concurrent_requests,
             timeout: Duration::from_secs(*timeout),
-        }
+
+            objectstore_client,
+            attachments_usecase,
+        }))
     }
 
     fn handle_message(&mut self, message: Upload) {
@@ -148,8 +170,17 @@ impl UploadService {
             return;
         }
 
+        let AttachmentScope {
+            organization_id,
+            project_id,
+        } = attachment.meta.scope;
+        let session = self
+            .attachments_usecase
+            .for_project(organization_id, project_id)
+            .session(&self.objectstore_client);
+
         self.pending_requests
-            .push(handle_upload(self.timeout, attachment, respond_to).boxed());
+            .push(handle_upload(self.timeout, session, attachment, respond_to).boxed());
     }
 }
 
@@ -161,7 +192,7 @@ impl Service for UploadService {
         loop {
             relay_log::trace!("Upload loop iteration");
             tokio::select! {
-                //Bias towards handling responses so that there's space for new incoming requests.
+                // Bias towards handling responses so that there's space for new incoming requests.
                 biased;
 
                 Some(result) = self.pending_requests.next() => {
@@ -197,26 +228,36 @@ fn count_upload(result: Result<(), Rejected<Error>>) {
 /// Returns an [`UploadedAttachment`] if the upload was successful.
 async fn handle_upload(
     timeout: Duration,
+    session: Result<Session, objectstore_client::Error>,
     attachment: Managed<Attachment>,
     respond_to: Recipient<Managed<UploadedAttachment>, NoResponse>,
 ) -> Result<(), Rejected<Error>> {
+    let session = session.map_err(|_err| attachment.reject_err(Error::UploadFailed))?;
+
     relay_log::trace!("Starting upload");
     let key = attachment.meta.attachment_id.as_deref();
-    let new_key = tokio::time::timeout(timeout, async { upload(key, &attachment.payload).await })
+
+    let mut put_request = session.put(attachment.payload.clone());
+    if let Some(key) = key {
+        put_request = put_request.key(key);
+    }
+    let future = async {
+        let result = put_request.send().await?;
+        Ok(result.key)
+    };
+
+    let new_key = tokio::time::timeout(timeout, future)
         .await
         .map_err(|_elapsed| attachment.reject_err(Error::Timeout))?
-        .map_err(|_error| attachment.reject_err(Error::UploadFailed))?;
+        .map_err(|_error: objectstore_client::Error| attachment.reject_err(Error::UploadFailed))?;
 
     if key.is_some_and(|key| key != new_key) {
         return Err(attachment.reject_err(Error::KeyMismatch));
     }
 
-    let uploaded_attachment = attachment.map(|Attachment { mut meta, payload }, _| {
-        meta.attachment_id = Some(new_key);
-        UploadedAttachment {
-            meta,
-            uploaded_bytes: payload.len(),
-        }
+    let uploaded_attachment = attachment.map(|Attachment { payload, .. }, _| UploadedAttachment {
+        stored_id: new_key,
+        uploaded_bytes: payload.len(),
     });
 
     respond_to.send(uploaded_attachment);
@@ -224,19 +265,8 @@ async fn handle_upload(
     Ok(())
 }
 
-/// Uploads the payload to the objectstore and returns the key under which it is stored.
-///
-/// - If `key` is `None`, the objectstore will assign a key.
-/// - If `key` is not `None`, write to objectstore with the given key.
-async fn upload(key: Option<&str>, payload: &[u8]) -> Result<String, ()> {
-    // TODO: call objectstore
-    Ok(key.unwrap_or_default().to_owned())
-}
-
 #[cfg(test)]
 mod tests {
-    use relay_redis::redis::FromRedisValue;
-    use relay_system::Addr;
 
     use super::*;
 
@@ -245,14 +275,21 @@ mod tests {
         let upload_service = UploadService::new(&UploadServiceConfig {
             max_concurrent_requests: 2,
             timeout: 1,
+            objectstore_url: Some("http://127.0.0.1:8888/".into()),
         })
+        .unwrap()
+        .unwrap()
         .start_detached();
 
         let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel();
 
-        let (attachment, mut managed_handle) = Managed::for_test(Attachment {
+        let (attachment, mut _managed_handle) = Managed::for_test(Attachment {
             meta: AttachmentMeta {
                 attachment_id: Some("my_key".to_owned()),
+                scope: AttachmentScope {
+                    organization_id: 123,
+                    project_id: 456,
+                },
             },
             payload: Bytes::from("hello world"),
         })
@@ -263,7 +300,7 @@ mod tests {
             respond_to: Recipient::<_, NoResponse>::new(tx),
         });
         let uploaded = rx.recv().await.unwrap();
-        assert_eq!(uploaded.meta.attachment_id.as_deref(), Some("my_key"));
+        assert_eq!(uploaded.stored_id, "my_key");
         uploaded.accept(|_| {});
 
         drop(upload_service);